refactor: Compatibility provision grouping
